+++ athena-scripts/raw_stage_optimisation_solution/scripts/raw_to_stage_etl/strategies/BackfillStrategy.py
@@ -1,11 +1,10 @@
 """BackfillStrategy is run after ScheduledStrategy to pick up records not picked up by ScheduledStrategy."""
 
-from raw_to_stage_etl.strategies.Strategy import Strategy
-from raw_to_stage_etl.util.exceptions.UtilExceptions import OperationFailedException
-
-from ..util.processing_utilities import (get_all_processed_dts, get_all_processed_times_per_day,
-                                         get_last_processed_time, get_min_timestamp_from_previous_run,
-                                         get_penultimate_processed_dt)
+from ..util.database_utilities import (get_all_processed_dts, get_all_processed_times_per_day,
+                                       get_min_timestamp_from_previous_run)
+from ..util.exceptions.UtilExceptions import OperationFailedException
+from ..util.pandas_dataframe_utilities import get_last_processed_time, get_penultimate_processed_dt
+from .Strategy import Strategy
 
 
 def get_raw_sql(
