+++ athena-scripts/raw_stage_optimisation_solution/scripts/raw_to_stage_etl/strategies/CustomStrategy.py
@@ -1,6 +1,6 @@
 """CustomStrategy is for ETL which can be run when required."""
 
-from ..util.processing_utilities import extract_element_by_name
+from ..util.json_config_processing_utilities import extract_element_by_name
 from .Strategy import Strategy
