+++ athena-scripts/raw_stage_optimisation_solution/scripts/raw_to_stage_etl/strategies/ScheduledStrategy.py
@@ -1,6 +1,7 @@
 """ScheduledStrategy is for ETL which runs everyday on schedule."""
 
-from ..util.processing_utilities import extract_element_by_name, get_max_processed_dt, get_max_timestamp
+from ..util.database_utilities import get_max_processed_dt, get_max_timestamp
+from ..util.json_config_processing_utilities import extract_element_by_name
 from .Strategy import Strategy
